changes the whole structure from class library to function library
