+++ examples/assets/csv/bird_env.csv
@@ -1,0 +1,1 @@
+0.02191,0.055385,0.10834,0.10682,0.032468,0.11197,0.029655,0.064478,0.093917,0.046128,0.058035,0.029161,0.02187,0.080961,0.11141,0.064204,0.10481,0.048962,0.12306,0.12162,0.27989,0.05193,0.03304,0.058248,0.13651,0.029007,0.022064,0.016962,0.10485,0.058471,0.051891,0.012875,0.0275,0.015494,0.13458,0.15881,0.33795,0.41067,0.35245,0.050953,0.31195,0.15969,0.59804,0.30414,0.44682,0.34788,0.20071,0.097253,0.30018,0.30492,0.15412,0.21509,0.27516,0.097603,0.022141,0.07769,0.054525,0.080975,0.11235,0.054015,0.017596,0.016517,0.030988,0.062983,0.085369,0.012393,0.011214,0.10905,0.024209,0.10282,0.040516,0.085029,0.034401,0.068537,0.042817,0.08125,0.084048,0.032298,0.0346,0.059535,0.17437,0.37664,0.17631,0.19767,0.28733,0.21191,0.44214,0.76366,0.19909,0.50599,0.17331,0.32526,0.25405,0.89816,0.027654,0.13715,0.14066,0.12816,0.15065,0.025923,0.18362,0.15763,0.095494,0.032345,0.1414,0.12575,0.052249,0.025065,0.066475,0.029915,0.022657,0.079852,0.063187,0.019824,0.025407,0.068101,0.045796,0.052538,0.053486,0.081918,0.0034909,0.039262,0.15251,0.028995,0.44154,0.68308,0.47945,0.6338,0.22793,0.18073,0.37373,0.51558,0.30821,0.33973,0.15672,0.16945,0.33672,0.61733,0.28205,0.048898,0.10046,0.10004,0.06818,0.021486,0.057909,0.043533,0.096026,0.035521,0.17019,0.052919,0.25424,0.19575,0.70978,0.31484,0.25311,0.45043,0.088596,0.42151,0.23836,0.074044,0.50164,0.091721,0.20167,0.17073,0.40316,0.15321,1,0.1682,0.39747,0.29036,0.12867,0.046868,0.052936,0.33107,0.111,0.13767,0.1221,0.30078,0.32347,0.087267,0.14278,0.063127,0.16324,0.048316,0.12491,0.060571,0.046236,0.034492,0.054987,0.031122,0.13665,0.076779,0.050757,0.014094,0.1062,0.077091,0.069546,0.034698,0.059358,0.084069,0.050716,0.092562,0.14807,0.11848,0.11385,0.075823,0.056022,0.050667,0.052161,0.05778,0.054435,0.074093,0.028544,0.091746,0.064632,0.063958,0.049158,0.028077,0.07124,0.10576,0.090834,0.045309,0.076662,0.04029,0.10722,0.030828,0.083467,0.17839,0.11843,0.15596,0.039907,0.20732,0.069012,0.047867,0.25258,0.085239,0.13594,0.019133,0.10404,0.076795,0.30991,0.16951,0.1423,0.26752,0.20886,0.60907,0.56896,0.54245,0.74046,0.27196,0.26214,0.16168,0.3455,0.15105,0.010803,0.27164,0.82988,0.50962,0.36366,0.33681,0.26877,0.11084,0.10646,0.029809,0.10858,0.076081,0.068637,0.013244,0.11376,0.075865,0.087615,0.07361,0.023311,0.24137,0.11999,0.11988,0.06308,0.12197,0.15552,0.18065,0.15018,0.14781,0.12467,0.22534,0.20917,0.56431,0.13115,0.3355,0.48964,0.14698,0.019058,0.22064,0.12573,0.16548,0.094562,0.15011,0.15337,0.12453,0.080592,0.11823,0.0085783,0.025163,0.080907,0.13301,0.076595,0.080817,0.034842,0.05923,0.11065,0.0066749,0.093727,0.058735,0.054824,0.017681,0.022653,0.07925,0.069861,0.047494,0.016713,0.14062,0.23995,0.3298,0.4273,0.4374,0.18867,0.4959,0.33852,0.38943,0.088453,0.42272,0.25907,0.87757,0.73281,0.52612,0.46316,0.17606,0.59983,0.46697,0.18282,0.3875,0.11895,0.069697,0.21087,0.14108,0.12605,0.10422,0.063949,0.11884,0.22533,0.34317,0.19993,0.051926,0.14781,0.097825,0.13033,0.17124,0.088652,0.071621,0.062288,0.17184,0.079553,0.094183,0.20355,0.13493,0.029246,0.30174,0.20957,0.44987,0.03279,0.21723,0.23707,0.15342,0.016125,0.079318,0.046773,0.13513,0.12803,0.059112,0.060071,0.084647,0.032432,0.04828,0.095886,0.17941,0.43899,0.32459,0.49123,0.11956,0.24746,0.29281,0.26517,0.59299,0.070832,0.3548,0.12159,0.13595,0.21888,0.13839,0.28092,0.03722,0.15243,0.10711,0.21013,0.15255,0.016099,0.18537,0.31312,0.45904,0.72154,0.33373,0.37975,0.35786,0.18497,0.20964,0.21259,0.27172,0.17901,0.048533,0.20604,0.33942,0.28506,0.18952,0.18818,0.23307,0.067825,0.14407,0.12665,0.47541,0.50111,0.41157,0.5626,0.27701,0.1226,0.27582,0.2003,0.42646,0.13792,0.066029,0.076979,0.051118,0.097147,0.091101,0.085884,0.12096,0.080078,0.060413,0.077711,0.083929,0.047942,0.14246,0.060005,0.065809,0.05908,0.021831,0.02336,0.12074,0.055363,0.069263,0.046305,0.074341,0.052324,0.065964,0.11043,0.058157,0.046267,0.061428,0.034187,0.0075864,0
