+++ examples/assets/csv/bird_timecode.csv
@@ -1,0 +1,1 @@
+0.0052083,0.015625,0.026042,0.036458,0.046875,0.057292,0.067708,0.078125,0.088542,0.098958,0.10937,0.11979,0.13021,0.14062,0.15104,0.16146,0.17188,0.18229,0.19271,0.20312,0.21354,0.22396,0.23438,0.24479,0.25521,0.26562,0.27604,0.28646,0.29687,0.30729,0.31771,0.32812,0.33854,0.34896,0.35937,0.36979,0.38021,0.39062,0.40104,0.41146,0.42187,0.43229,0.44271,0.45312,0.46354,0.47396,0.48437,0.49479,0.50521,0.51562,0.52604,0.53646,0.54688,0.55729,0.56771,0.57812,0.58854,0.59896,0.60938,0.61979,0.63021,0.64062,0.65104,0.66146,0.67188,0.68229,0.69271,0.70312,0.71354,0.72396,0.73438,0.74479,0.75521,0.76562,0.77604,0.78646,0.79688,0.80729,0.81771,0.82812,0.83854,0.84896,0.85938,0.86979,0.88021,0.89062,0.90104,0.91146,0.92188,0.93229,0.94271,0.95312,0.96354,0.97396,0.98438,0.99479,1.0052,1.0156,1.026,1.0365,1.0469,1.0573,1.0677,1.0781,1.0885,1.099,1.1094,1.1198,1.1302,1.1406,1.151,1.1615,1.1719,1.1823,1.1927,1.2031,1.2135,1.224,1.2344,1.2448,1.2552,1.2656,1.276,1.2865,1.2969,1.3073,1.3177,1.3281,1.3385,1.349,1.3594,1.3698,1.3802,1.3906,1.401,1.4115,1.4219,1.4323,1.4427,1.4531,1.4635,1.474,1.4844,1.4948,1.5052,1.5156,1.526,1.5365,1.5469,1.5573,1.5677,1.5781,1.5885,1.599,1.6094,1.6198,1.6302,1.6406,1.651,1.6615,1.6719,1.6823,1.6927,1.7031,1.7135,1.724,1.7344,1.7448,1.7552,1.7656,1.776,1.7865,1.7969,1.8073,1.8177,1.8281,1.8385,1.849,1.8594,1.8698,1.8802,1.8906,1.901,1.9115,1.9219,1.9323,1.9427,1.9531,1.9635,1.974,1.9844,1.9948,2.0052,2.0156,2.026,2.0365,2.0469,2.0573,2.0677,2.0781,2.0885,2.099,2.1094,2.1198,2.1302,2.1406,2.151,2.1615,2.1719,2.1823,2.1927,2.2031,2.2135,2.224,2.2344,2.2448,2.2552,2.2656,2.276,2.2865,2.2969,2.3073,2.3177,2.3281,2.3385,2.349,2.3594,2.3698,2.3802,2.3906,2.401,2.4115,2.4219,2.4323,2.4427,2.4531,2.4635,2.474,2.4844,2.4948,2.5052,2.5156,2.526,2.5365,2.5469,2.5573,2.5677,2.5781,2.5885,2.599,2.6094,2.6198,2.6302,2.6406,2.651,2.6615,2.6719,2.6823,2.6927,2.7031,2.7135,2.724,2.7344,2.7448,2.7552,2.7656,2.776,2.7865,2.7969,2.8073,2.8177,2.8281,2.8385,2.849,2.8594,2.8698,2.8802,2.8906,2.901,2.9115,2.9219,2.9323,2.9427,2.9531,2.9635,2.974,2.9844,2.9948,3.0052,3.0156,3.026,3.0365,3.0469,3.0573,3.0677,3.0781,3.0885,3.099,3.1094,3.1198,3.1302,3.1406,3.151,3.1615,3.1719,3.1823,3.1927,3.2031,3.2135,3.224,3.2344,3.2448,3.2552,3.2656,3.276,3.2865,3.2969,3.3073,3.3177,3.3281,3.3385,3.349,3.3594,3.3698,3.3802,3.3906,3.401,3.4115,3.4219,3.4323,3.4427,3.4531,3.4635,3.474,3.4844,3.4948,3.5052,3.5156,3.526,3.5365,3.5469,3.5573,3.5677,3.5781,3.5885,3.599,3.6094,3.6198,3.6302,3.6406,3.651,3.6615,3.6719,3.6823,3.6927,3.7031,3.7135,3.724,3.7344,3.7448,3.7552,3.7656,3.776,3.7865,3.7969,3.8073,3.8177,3.8281,3.8385,3.849,3.8594,3.8698,3.8802,3.8906,3.901,3.9115,3.9219,3.9323,3.9427,3.9531,3.9635,3.974,3.9844,3.9948,4.0052,4.0156,4.026,4.0365,4.0469,4.0573,4.0677,4.0781,4.0885,4.099,4.1094,4.1198,4.1302,4.1406,4.151,4.1615,4.1719,4.1823,4.1927,4.2031,4.2135,4.224,4.2344,4.2448,4.2552,4.2656,4.276,4.2865,4.2969,4.3073,4.3177,4.3281,4.3385,4.349,4.3594,4.3698,4.3802,4.3906,4.401,4.4115,4.4219,4.4323,4.4427,4.4531,4.4635,4.474,4.4844,4.4948,4.5052,4.5156,4.526,4.5365,4.5469,4.5573,4.5677,4.5781,4.5885,4.599,4.6094,4.6198,4.6302,4.6406,4.651,4.6615,4.6719,4.6823,4.6927,4.7031,4.7135,4.724,4.7344,4.7448,4.7552,4.7656,4.776,4.7865,4.7969,4.8073,4.8177,4.8281,4.8385,4.849,4.8594,4.8698,4.8802
