+++ index.js
@@ -1,3 +1,2 @@
-const S2M = require('./lib/S2M');
-
-module.exports = S2M;
+const s2m = require('./lib');
+module.exports = s2m;
