+++ lib/index.js
@@ -1,0 +1,14 @@
+const utils = require('./utils');
+const txtImport = require('./import');
+const partials2melodies = require('./partials2melodies');
+const genSMFs = require('./genSMFs');
+const {smfExport, smfsBatchExport} = require('./export');
+
+module.exports = {
+    utils, 
+    txtImport, 
+    partials2melodies,
+    genSMFs,
+    smfExport,
+    smfsBatchExport
+};
