--- lib/viewer/index.js
@@ -1,7 +1,0 @@
-const { app, BrowserWindow } = require('electron');
-
-app.on('ready', () => {
-    const window = new BrowserWindow({ width: 800, height: 600 });
-
-    window.loadFile('./viewer.html');
-});
