--- lib/viewer/viewer.html
@@ -1,21 +1,0 @@
-<!doctype html>
-<html>
-  <head>
-    <title>plot for debugging</title>
-    <script src="https://cdn.plot.ly/plotly-latest.min.js"></script>
-    <style>
-        html, body {
-        height: 100%;
-        margin: 0;
-        }
-
-        .full-height {
-        height: 100%;
-        }
-    </style>
-  </head>
-  <body>
-    <div id="graph" class="full-height"></div>
-    <script src="./viewer.js"></script>
-  </body>
-</html>
