+++ package.json
@@ -7,7 +7,7 @@
     "lib": "lib"
   },
   "scripts": {
-    "test": "mocha"
+    "test": "mocha --timeout 15000"
   },
   "dependencies": {
     "chai": "^4.2.0",
@@ -16,6 +16,7 @@
     "jzz": "^0.5.7",
     "jzz-midi-smf": "^1.0.3",
     "lodash": "^4.17.11",
+    "mime-types": "^2.1.21",
     "mocha": "^5.2.0",
     "moment": "^2.22.2",
     "node-osc": "^2.1.0",
