--- test/.gitignore
@@ -1,2 +1,0 @@
-assets/archive
-assets/wav
